Moved back functions from synchapi.py to kernel32.py. Updated CREDITS.TXT
