+++ CREDITS.TXT
@@ -48,3 +48,4 @@ iamyeh
 alfink
 bambu
 kabeor
+0ssigeno
