+++ qiling/os/windows/dlls/__init__.py
@@ -11,4 +11,3 @@ from .msi import *
 from .ntdll import *
 from .ucrtbased import *
 from .wsock32 import *
-from .synchapi import *
