--- qiling/os/windows/dlls/synchapi.py
@@ -1,61 +1,0 @@
-#!/usr/bin/env python3
-#
-# Cross Platform and Multi Architecture Advanced Binary Emulation Framework
-# Built on top of Unicorn emulator (www.unicorn-engine.org)
-
-import struct
-import time
-from qiling.os.windows.const import *
-from qiling.os.fncc import *
-from qiling.os.windows.fncc import *
-from qiling.os.windows.utils import *
-from qiling.os.memory import align
-from qiling.os.windows.thread import *
-from qiling.os.windows.handle import *
-from qiling.exception import *
-
-
-# void EnterCriticalSection(
-#  LPCRITICAL_SECTION lpCriticalSection
-# );
-@winapi(cc=STDCALL, params={
-    "lpCriticalSection": POINTER
-})
-def hook_EnterCriticalSection(ql, address, params):
-    return 0
-
-
-# void LeaveCriticalSection(
-#  LPCRITICAL_SECTION lpCriticalSection
-# );
-@winapi(cc=STDCALL, params={
-    "lpCriticalSection": POINTER
-})
-def hook_LeaveCriticalSection(ql, address, params):
-    return 0
-
-
-# BOOL InitializeCriticalSectionEx(
-#   LPCRITICAL_SECTION lpCriticalSection,
-#   DWORD              dwSpinCount,
-#   DWORD              Flags
-# );
-@winapi(cc=STDCALL, params={
-    "lpCriticalSection": POINTER,
-    "dwSpinCount": DWORD,
-    "Flags": DWORD
-})
-def hook_InitializeCriticalSectionEx(ql, address, params):
-    return 1
-
-
-# BOOL InitializeCriticalSectionAndSpinCount(
-#  LPCRITICAL_SECTION lpCriticalSection,
-#  DWORD              dwSpinCount
-# );
-@winapi(cc=STDCALL, params={
-    "lpCriticalSection": POINTER,
-    "dwSpinCount": UINT
-})
-def hook_InitializeCriticalSectionAndSpinCount(ql, address, params):
-    return 1
